+++ test/clipboard_test.dart
@@ -5,13 +5,23 @@ import 'dart:ffi';
 
 import 'package:checks/checks.dart';
 import 'package:ffi/ffi.dart';
+import 'package:ffi_leak_tracker/ffi_leak_tracker.dart';
 import 'package:test/scaffolding.dart';
 import 'package:win32/win32.dart';
 import 'package:win32_clipboard/win32_clipboard.dart';
 
 void main() {
   group('Clipboard', () {
-    setUp(Clipboard.clear);
+    setUpAll(LeakTracker.enableInDebug);
+
+    setUp(() {
+      Clipboard.clear();
+      LeakTracker.reset();
+    });
+
+    tearDown(LeakTracker.verifyNoLeaksInDebug);
+
+    tearDownAll(Clipboard.clear);
 
     test('clear', () {
       check(Clipboard.setText('win32_clipboard')).isTrue();
@@ -26,8 +36,9 @@ void main() {
         check(format.id).not((it) => it.equals(0));
         check(format.name).equals('win32_clipboard PointerData');
         final ptr = calloc<Uint8>()..value = 10;
-        check(Clipboard.setData(ClipboardData.pointer(ptr, 1, format)))
-            .isTrue();
+        check(
+          Clipboard.setData(ClipboardData.pointer(ptr, 1, format)),
+        ).isTrue();
         free(ptr);
         check(Clipboard.formats).deepEquals([format]);
       });
@@ -35,24 +46,25 @@ void main() {
       test('UnicodeText', () {
         check(Clipboard.formats).isEmpty();
         check(Clipboard.setText('🌍 win32_clipboard 🚀')).isTrue();
-        check(Clipboard.formats).which((it) => it
-          ..length.isGreaterOrEqual(4)
-          ..contains(
-              const ClipboardFormat(CLIPBOARD_FORMAT.CF_LOCALE, 'CF_LOCALE'))
-          ..contains(
-              const ClipboardFormat(CLIPBOARD_FORMAT.CF_OEMTEXT, 'CF_OEMTEXT'))
-          ..contains(const ClipboardFormat(CLIPBOARD_FORMAT.CF_TEXT, 'CF_TEXT'))
-          ..contains(ClipboardFormat.unicodeText));
+        check(Clipboard.formats).which(
+          (it) => it
+            ..length.isGreaterOrEqual(4)
+            ..contains(const .new(CF_LOCALE, 'CF_LOCALE'))
+            ..contains(const .new(CF_OEMTEXT, 'CF_OEMTEXT'))
+            ..contains(const .new(CF_TEXT, 'CF_TEXT'))
+            ..contains(.unicodeText),
+        );
       });
     });
 
     group('getData', () {
       test('FileList', () {
-        check(Clipboard.setData(
-                const ClipboardData.fileList([r'c:\src\a.dart', r'd:\b.txt'])))
-            .isTrue();
-        check(Clipboard.getData(ClipboardFormat.fileList)).isNotNull().equals(
-            const ClipboardData.fileList([r'c:\src\a.dart', r'd:\b.txt']));
+        check(
+          Clipboard.setData(const .fileList([r'c:\src\a.dart', r'd:\b.txt'])),
+        ).isTrue();
+        check(
+          Clipboard.getData(.fileList),
+        ).isNotNull().equals(const .fileList([r'c:\src\a.dart', r'd:\b.txt']));
       });
 
       test('Pointer', () {
@@ -65,8 +77,7 @@ void main() {
           ..[2] = 30
           ..[3] = 40
           ..[4] = 50;
-        check(Clipboard.setData(ClipboardData.pointer(ptr, 5, format)))
-            .isTrue();
+        check(Clipboard.setData(.pointer(ptr, 5, format))).isTrue();
         free(ptr);
         final data = Clipboard.getData(format);
         check(data).isA<PointerData>();
@@ -77,23 +88,23 @@ void main() {
       });
 
       test('UnicodeText', () {
-        check(Clipboard.setData(
-                const ClipboardData.unicodeText('🌍 win32_clipboard 🚀')))
-            .isTrue();
-        check(Clipboard.getData(ClipboardFormat.unicodeText))
-            .isNotNull()
-            .equals(const ClipboardData.unicodeText('🌍 win32_clipboard 🚀'));
+        check(
+          Clipboard.setData(const .unicodeText('🌍 win32_clipboard 🚀')),
+        ).isTrue();
+        check(
+          Clipboard.getData(.unicodeText),
+        ).isNotNull().equals(const .unicodeText('🌍 win32_clipboard 🚀'));
       });
     });
 
     group('setData', () {
       test('FileList', () {
-        check(Clipboard.setData(
-                const ClipboardData.fileList([r'c:\src\a.dart', r'd:\b.txt'])))
-            .isTrue();
-        check(Clipboard.getData(ClipboardFormat.fileList))
-            .isNotNull()
-            .equals(const FileListData([r'c:\src\a.dart', r'd:\b.txt']));
+        check(
+          Clipboard.setData(const .fileList([r'c:\src\a.dart', r'd:\b.txt'])),
+        ).isTrue();
+        check(
+          Clipboard.getData(.fileList),
+        ).isNotNull().equals(const .fileList([r'c:\src\a.dart', r'd:\b.txt']));
       });
 
       test('Pointer', () {
@@ -106,8 +117,7 @@ void main() {
           ..[2] = 30
           ..[3] = 40
           ..[4] = 50;
-        check(Clipboard.setData(ClipboardData.pointer(ptr, 5, format)))
-            .isTrue();
+        check(Clipboard.setData(.pointer(ptr, 5, format))).isTrue();
         free(ptr);
         final data = Clipboard.getData(format);
         check(data).isA<PointerData>();
@@ -118,20 +128,21 @@ void main() {
       });
 
       test('UnicodeText', () {
-        check(Clipboard.setData(
-                const ClipboardData.unicodeText('🌍 win32_clipboard 🚀')))
-            .isTrue();
-        check(Clipboard.getData(ClipboardFormat.unicodeText))
-            .isNotNull()
-            .equals(const UnicodeTextData('🌍 win32_clipboard 🚀'));
+        check(
+          Clipboard.setData(const .unicodeText('🌍 win32_clipboard 🚀')),
+        ).isTrue();
+        check(
+          Clipboard.getData(.unicodeText),
+        ).isNotNull().equals(const .unicodeText('🌍 win32_clipboard 🚀'));
       });
     });
 
     test('getFileList', () {
       check(Clipboard.getFileList()).isEmpty();
       check(Clipboard.setFileList([r'c:\src\a.dart', r'd:\b.txt'])).isTrue();
-      check(Clipboard.getFileList())
-          .deepEquals([r'c:\src\a.dart', r'd:\b.txt']);
+      check(
+        Clipboard.getFileList(),
+      ).deepEquals([r'c:\src\a.dart', r'd:\b.txt']);
     });
 
     test('setFileList', () {
@@ -141,8 +152,9 @@ void main() {
           .has((e) => e.message, 'message')
           .equals('Cannot be empty');
       check(Clipboard.setFileList([r'c:\src\a.dart', r'd:\b.txt'])).isTrue();
-      check(Clipboard.getFileList())
-          .deepEquals([r'c:\src\a.dart', r'd:\b.txt']);
+      check(
+        Clipboard.getFileList(),
+      ).deepEquals([r'c:\src\a.dart', r'd:\b.txt']);
     });
 
     test('getText', () {
@@ -159,35 +171,28 @@ void main() {
     test('getPriorityFormat', () {
       const formats = [
         ClipboardFormat.unicodeText,
-        ClipboardFormat(CLIPBOARD_FORMAT.CF_TEXT, 'CF_TEXT'),
-        ClipboardFormat(CLIPBOARD_FORMAT.CF_LOCALE, 'CF_LOCALE'),
-        ClipboardFormat(CLIPBOARD_FORMAT.CF_OEMTEXT, 'CF_OEMTEXT'),
+        ClipboardFormat(CF_TEXT, 'CF_TEXT'),
+        ClipboardFormat(CF_LOCALE, 'CF_LOCALE'),
+        ClipboardFormat(CF_OEMTEXT, 'CF_OEMTEXT'),
       ];
       check(Clipboard.getPriorityFormat(formats)).isNull();
       check(Clipboard.setText('win32_clipboard')).isTrue();
-      check(Clipboard.getPriorityFormat(formats))
-          .isNotNull()
-          .equals(ClipboardFormat.unicodeText);
-      check(Clipboard.getPriorityFormat(formats.reversed.toList()))
-          .isNotNull()
-          .equals(
-              const ClipboardFormat(CLIPBOARD_FORMAT.CF_OEMTEXT, 'CF_OEMTEXT'));
+      check(
+        Clipboard.getPriorityFormat(formats),
+      ).isNotNull().equals(.unicodeText);
+      check(
+        Clipboard.getPriorityFormat(formats.reversed.toList()),
+      ).isNotNull().equals(const .new(CF_OEMTEXT, 'CF_OEMTEXT'));
     });
 
     test('hasFormat', () {
-      check(Clipboard.hasFormat(ClipboardFormat.unicodeText)).isFalse();
+      check(Clipboard.hasFormat(.unicodeText)).isFalse();
       check(Clipboard.setText('win32_clipboard')).isTrue();
-      check(Clipboard.hasFormat(
-              const ClipboardFormat(CLIPBOARD_FORMAT.CF_LOCALE, 'CF_LOCALE')))
-          .isTrue();
-      check(Clipboard.hasFormat(
-              const ClipboardFormat(CLIPBOARD_FORMAT.CF_OEMTEXT, 'CF_OEMTEXT')))
-          .isTrue();
-      check(Clipboard.hasFormat(
-              const ClipboardFormat(CLIPBOARD_FORMAT.CF_TEXT, 'CF_TEXT')))
-          .isTrue();
-      check(Clipboard.hasFormat(ClipboardFormat.unicodeText)).isTrue();
-      check(Clipboard.hasFormat(ClipboardFormat.fileList)).isFalse();
+      check(Clipboard.hasFormat(const .new(CF_LOCALE, 'CF_LOCALE'))).isTrue();
+      check(Clipboard.hasFormat(const .new(CF_OEMTEXT, 'CF_OEMTEXT'))).isTrue();
+      check(Clipboard.hasFormat(const .new(CF_TEXT, 'CF_TEXT'))).isTrue();
+      check(Clipboard.hasFormat(.unicodeText)).isTrue();
+      check(Clipboard.hasFormat(.fileList)).isFalse();
     });
 
     test('hasFileList', () {
@@ -215,149 +220,162 @@ void main() {
       check(Clipboard.numberOfFormats).isGreaterOrEqual(4);
     });
 
-    group('onDataChanged', () {
-      test('emits FileListData when file list data is placed on the clipboard',
-          () async {
-        final result = <FileListData>[];
-        final subscription =
-            Clipboard.onDataChanged(ClipboardFormat.fileList).listen((data) {
-          if (data is FileListData) result.add(data);
-        });
+    // group('onDataChanged', () {
+    //   test(
+    //     'emits FileListData when file list data is placed on the clipboard',
+    //     () async {
+    //       final result = <FileListData>[];
+    //       final subscription = Clipboard.onDataChanged(.fileList).listen((
+    //         data,
+    //       ) {
+    //         if (data is FileListData) result.add(data);
+    //       });
 
-        // Allow time for the subscription to be set up.
-        await Future.delayed(const Duration(milliseconds: 50));
+    //       // Allow time for the subscription to be set up.
+    //       await Future.delayed(const Duration(milliseconds: 50));
 
-        // Set some file list data to the clipboard to simulate a change.
-        const data = FileListData([r'c:\src\a.dart', r'd:\b.txt']);
-        check(Clipboard.setData(data)).isTrue();
+    //       // Set some file list data to the clipboard to simulate a change.
+    //       const data = FileListData([r'c:\src\a.dart', r'd:\b.txt']);
+    //       check(Clipboard.setData(data)).isTrue();
 
-        // Yield control to the Dart event loop to allow time for the clipboard
-        // change to be detected.
-        await Future.delayed(const Duration(milliseconds: 1));
+    //       // Yield control to the Dart event loop to allow time for the clipboard
+    //       // change to be detected.
+    //       await Future.delayed(const Duration(milliseconds: 1));
 
-        // Expect the stream to emit the FileListData.
-        check(result.isNotEmpty).isTrue();
-        check(result.first).equals(data);
+    //       // Expect the stream to emit the FileListData.
+    //       check(result.isNotEmpty).isTrue();
+    //       check(result.first).equals(data);
 
-        await subscription.cancel();
-      });
+    //       await subscription.cancel();
+    //     },
+    //   );
 
-      test('emits PointerData when pointer data is placed on the clipboard',
-          () async {
-        final format = Clipboard.registerFormat('win32_clipboard PointerData');
-        check(format.id).not((it) => it.equals(0));
-        final result = <PointerData>[];
-        final subscription = Clipboard.onDataChanged(format).listen((data) {
-          if (data is PointerData) result.add(data);
-        });
+    //   test(
+    //     'emits PointerData when pointer data is placed on the clipboard',
+    //     () async {
+    //       final format = Clipboard.registerFormat(
+    //         'win32_clipboard PointerData',
+    //       );
+    //       check(format.id).not((it) => it.equals(0));
+    //       final result = <PointerData>[];
+    //       final subscription = Clipboard.onDataChanged(format).listen((data) {
+    //         if (data is PointerData) result.add(data);
+    //       });
 
-        // Allow time for the subscription to be set up.
-        await Future.delayed(const Duration(milliseconds: 50));
+    //       // Allow time for the subscription to be set up.
+    //       await Future.delayed(const Duration(milliseconds: 50));
 
-        // Set some pointer data to the clipboard to simulate a change.
-        final ptr = calloc<Uint8>()..value = 10;
-        final data = PointerData(ptr, 1, format);
-        check(Clipboard.setData(data)).isTrue();
-        free(ptr);
+    //       // Set some pointer data to the clipboard to simulate a change.
+    //       final ptr = calloc<Uint8>()..value = 10;
+    //       final data = PointerData(ptr, 1, format);
+    //       check(Clipboard.setData(data)).isTrue();
+    //       free(ptr);
 
-        // Yield control to the Dart event loop to allow time for the clipboard
-        // change to be detected.
-        await Future.delayed(const Duration(milliseconds: 1));
+    //       // Yield control to the Dart event loop to allow time for the clipboard
+    //       // change to be detected.
+    //       await Future.delayed(const Duration(milliseconds: 1));
 
-        // Expect the stream to emit the PointerData.
-        check(result.isNotEmpty).isTrue();
-        check(result.first)
-          ..has((p) => p.pointer, 'pointer')
-              .has((p) => p.address, 'address')
-              .not((it) => it.equals(0))
-          ..has((p) => p.lengthInBytes, 'lengthInBytes').equals(1)
-          ..has((p) => p.format, 'format').equals(format);
+    //       // Expect the stream to emit the PointerData.
+    //       check(result.isNotEmpty).isTrue();
+    //       check(result.first)
+    //         ..has(
+    //           (p) => p.pointer,
+    //           'pointer',
+    //         ).has((p) => p.address, 'address').not((it) => it.equals(0))
+    //         ..has((p) => p.lengthInBytes, 'lengthInBytes').equals(1)
+    //         ..has((p) => p.format, 'format').equals(format);
 
-        await subscription.cancel();
-      });
+    //       await subscription.cancel();
+    //     },
+    //   );
 
-      test('emits UnicodeTextData when text data is placed on the clipboard',
-          () async {
-        final result = <UnicodeTextData>[];
-        final subscription =
-            Clipboard.onDataChanged(ClipboardFormat.unicodeText).listen((data) {
-          if (data is UnicodeTextData) result.add(data);
-        });
+    //   test(
+    //     'emits UnicodeTextData when text data is placed on the clipboard',
+    //     () async {
+    //       final result = <UnicodeTextData>[];
+    //       final subscription = Clipboard.onDataChanged(.unicodeText).listen((
+    //         data,
+    //       ) {
+    //         if (data is UnicodeTextData) result.add(data);
+    //       });
 
-        // Allow time for the subscription to be set up.
-        await Future.delayed(const Duration(milliseconds: 50));
+    //       // Allow time for the subscription to be set up.
+    //       await Future.delayed(const Duration(milliseconds: 50));
 
-        // Set some text data to the clipboard to simulate a change.
-        const data = UnicodeTextData('Hello, Clipboard!');
-        check(Clipboard.setData(data)).isTrue();
+    //       // Set some text data to the clipboard to simulate a change.
+    //       const data = UnicodeTextData('Hello, Clipboard!');
+    //       check(Clipboard.setData(data)).isTrue();
 
-        // Yield control to the Dart event loop to allow time for the clipboard
-        // change to be detected.
-        await Future.delayed(const Duration(milliseconds: 1));
+    //       // Yield control to the Dart event loop to allow time for the clipboard
+    //       // change to be detected.
+    //       await Future.delayed(const Duration(milliseconds: 1));
 
-        // Expect the stream to emit the UnicodeTextData.
-        check(result.isNotEmpty).isTrue();
-        check(result.first).equals(data);
+    //       // Expect the stream to emit the UnicodeTextData.
+    //       check(result.isNotEmpty).isTrue();
+    //       check(result.first).equals(data);
 
-        await subscription.cancel();
-      });
-    });
+    //       await subscription.cancel();
+    //     },
+    //   );
+    // });
 
-    group('onFileListChanged', () {
-      test('emits file list when file list data is placed on the clipboard',
-          () async {
-        final result = <List<String>>[];
-        final subscription = Clipboard.onFileListChanged.listen(result.add);
+    // group('onFileListChanged', () {
+    //   test(
+    //     'emits file list when file list data is placed on the clipboard',
+    //     () async {
+    //       final result = <List<String>>[];
+    //       final subscription = Clipboard.onFileListChanged.listen(result.add);
 
-        // Allow time for the subscription to be set up.
-        await Future.delayed(const Duration(milliseconds: 50));
+    //       // Allow time for the subscription to be set up.
+    //       await Future.delayed(const Duration(milliseconds: 50));
 
-        // Set some file list data to the clipboard to simulate a change.
-        const files = [r'c:\src\a.dart', r'd:\b.txt'];
-        check(Clipboard.setFileList(files)).isTrue();
+    //       // Set some file list data to the clipboard to simulate a change.
+    //       const files = [r'c:\src\a.dart', r'd:\b.txt'];
+    //       check(Clipboard.setFileList(files)).isTrue();
 
-        // Yield control to the Dart event loop to allow time for the clipboard
-        // change to be detected.
-        await Future.delayed(const Duration(milliseconds: 1));
+    //       // Yield control to the Dart event loop to allow time for the clipboard
+    //       // change to be detected.
+    //       await Future.delayed(const Duration(milliseconds: 1));
 
-        // Expect the stream to emit the file list data.
-        check(result.isNotEmpty).isTrue();
-        check(result.first).deepEquals(files);
+    //       // Expect the stream to emit the file list data.
+    //       check(result.isNotEmpty).isTrue();
+    //       check(result.first).deepEquals(files);
 
-        await subscription.cancel();
-      });
-    });
+    //       await subscription.cancel();
+    //     },
+    //   );
+    // });
 
-    group('onTextChanged', () {
-      test('emits string when text data is placed on the clipboard', () async {
-        final result = <String>[];
-        final subscription = Clipboard.onTextChanged.listen(result.add);
+    // group('onTextChanged', () {
+    //   test('emits string when text data is placed on the clipboard', () async {
+    //     final result = <String>[];
+    //     final subscription = Clipboard.onTextChanged.listen(result.add);
 
-        // Allow time for the subscription to be set up.
-        await Future.delayed(const Duration(milliseconds: 50));
+    //     // Allow time for the subscription to be set up.
+    //     await Future.delayed(const Duration(milliseconds: 50));
 
-        // Set some text data to the clipboard to simulate a change.
-        const text = 'Hello, Clipboard!';
-        check(Clipboard.setText(text)).isTrue();
+    //     // Set some text data to the clipboard to simulate a change.
+    //     const text = 'Hello, Clipboard!';
+    //     check(Clipboard.setText(text)).isTrue();
 
-        // Yield control to the Dart event loop to allow time for the clipboard
-        // change to be detected.
-        await Future.delayed(const Duration(milliseconds: 1));
+    //     // Yield control to the Dart event loop to allow time for the clipboard
+    //     // change to be detected.
+    //     await Future.delayed(const Duration(milliseconds: 1));
 
-        // Expect the stream to emit the text data.
-        check(result.isNotEmpty).isTrue();
-        check(result.first).equals(text);
+    //     // Expect the stream to emit the text data.
+    //     check(result.isNotEmpty).isTrue();
+    //     check(result.first).equals(text);
 
-        await subscription.cancel();
-      });
-    });
+    //     await subscription.cancel();
+    //   });
+    // });
 
     test('registerFormat', () {
       final format = Clipboard.registerFormat('win32_clipboard PointerData');
       check(format.id).not((it) => it.equals(0));
       check(format.name).equals('win32_clipboard PointerData');
       final ptr = calloc<Uint8>()..value = 10;
-      check(Clipboard.setData(ClipboardData.pointer(ptr, 1, format))).isTrue();
+      check(Clipboard.setData(.pointer(ptr, 1, format))).isTrue();
       free(ptr);
       check(Clipboard.hasFormat(format)).isTrue();
       final data = Clipboard.getData(format);
@@ -367,7 +385,5 @@ void main() {
       final bytes = pointer.asTypedList(lengthInBytes);
       check(bytes).deepEquals([10]);
     });
-
-    tearDownAll(Clipboard.clear);
   });
 }