+++ test/format_test.dart
@@ -7,14 +7,14 @@ void main() {
   group('ClipboardFormat', () {
     test('fileList', () {
       const format = ClipboardFormat.fileList;
-      check(format.id).equals(CLIPBOARD_FORMAT.CF_HDROP);
+      check(format.id).equals(CF_HDROP);
       check(format.name).equals('CF_HDROP');
       check(format.toString()).equals('ClipboardFormat(15, CF_HDROP)');
     });
 
     test('unicodeText', () {
       const format = ClipboardFormat.unicodeText;
-      check(format.id).equals(CLIPBOARD_FORMAT.CF_UNICODETEXT);
+      check(format.id).equals(CF_UNICODETEXT);
       check(format.name).equals('CF_UNICODETEXT');
       check(format.toString()).equals('ClipboardFormat(13, CF_UNICODETEXT)');
     });